1 working test, server.js created and seperated from index.js. supertest and jest installed
