+++ knexfile.js
@@ -29,6 +29,25 @@ module.exports = {
       directory: "./database/seeds",
     },
     },
+    testing: {
+      client: 'sqlite3',
+      connection: {
+        filename: './database/test.db3',
+      },
+      useNullAsDefault: true,
+      migrations: {
+        directory: './database/migrations',
+      },
+      seeds: {
+        directory: './database/seeds',
+      },
+      pool: {
+        afterCreate: (conn, done) => {
+          conn.run('PRAGMA foreign_keys = ON', done)
+        },
+      },
+    },
   }
+
