+++ package.json
@@ -1,6 +1,7 @@
 {
 	"main": "index.js",
 	"scripts": {
+		"test": "cross-env DB_ENV=testing jest --watch",
 		"server": "nodemon -r dotenv/config index.js",
 		"start": "node index.js"
 	},
@@ -9,7 +10,9 @@
 	},
 	"devDependencies": {
 		"dotenv": "^8.2.0",
-		"nodemon": "^2.0.4"
+		"jest": "^26.1.0",
+		"nodemon": "^2.0.4",
+		"supertest": "^4.0.2"
 	},
 	"dependencies": {
 		"bcryptjs": "^2.4.3",
